refactor(schemas): migrate Phase 6 endpoint schemas to domain schema files (#6042)
Move locally-defined Pydantic models from 3 endpoint files into domain schema files:
- schemas_analytics.py: Cost/Budget schemas (analytics_cost.py)
- schemas_knowledge.py: Memory graph schemas (memory.py)
- schemas_agent.py: Team management schemas (user_management/teams.py)
